### 2021-06-14 #3365 @whedon
[PRE REVIEW]: pointcloudset: Efficient analysis of large datasets of point clouds recorded over time
**Submitting author:** @tgoelles (<a href="http://orcid.org/0000-0002-3925-6260">Thomas Goelles</a>)
**Repository:** <a href="https://github.com/virtual-vehicle/pointcloudset/" target ="_blank">https://github.com/virtual-vehicle/pointcloudset/</a>
**Version:** 0.2.2
