### 2021-06-15 560c9dd
delete envronment.yml
