+++ paper/paper.bib
@@ -121,6 +121,7 @@
   booktitle = { Proceedings of the 14th Python in Science Conference },
   pages     = { 130 - 136 },
   year      = { 2015 },
-  editor    = { Kathryn Huff and James Bergstra }
+  editor    = { Kathryn Huff and James Bergstra },
+  doi = {10.25080/majora-7b98e3ed-013}
 }
 
