### 2021-07-01 93ef56a
added basic checks for dir
