### 2021-07-01 115ac0d
added basic dataset dir checks
