### 2021-07-01 415a5f9
deleted old comments in docker file
