+++ .github/workflows/doc.yml
@@ -17,32 +17,22 @@ jobs:
       - uses: actions/checkout@v2
         with:
           fetch-depth: 0
-      - name: Cache conda
-        uses: actions/cache@v2
-        env:
-          # Increase this value to reset cache if etc/example-environment.yml has not changed
-          CACHE_NUMBER: 0
+
+      - name: Setup Conda Environment
+        uses: conda-incubator/setup-miniconda@v2
         with:
-          path: ~/conda_pkgs_dir
-          key: ${{ runner.os }}-conda-${{ env.CACHE_NUMBER }}-${{
-            hashFiles('etc/example-environment.yml') }}
-      - uses: conda-incubator/setup-miniconda@v2
-        with:
-          activate-environment: test
-          auto-activate-base: false
-          environment-file: docker/environment.yml
-          use-only-tar-bz2: true # IMPORTANT: This needs to be set for caching to work properly!
-      - name: test conda
-        shell: bash -l {0}
-        run: |
-          pip install -e .
-          conda info
-          conda list
+          miniconda-version: "latest"
+          environment-file: docker/environment.yaml
+          activate-environment: base
+          auto-activate-base: true
+
       - name: make doc
         shell: bash -l {0}
         run: make doc
+
       - name: add .nojekyll
         run: touch doc/sphinx/build/html/.nojekyll
+
       - name: Deploy to GitHub Pages
         if: success()
         uses: crazy-max/ghaction-github-pages@v2
