+++ doc/sphinx/ci-environment.yml
@@ -27,7 +27,7 @@ dependencies:
       - mypy
       - pylint
       - flake8
-      - pandoc
+      - pandoc>2.0
       - typer
       - nbval
       - geopandas
