### 2021-07-01 d08dc44
better error massage with datset check
