+++ pointcloudset/dataset_core.py
@@ -157,7 +157,7 @@ class DatasetCore:
         if len(self) > 0:
             assert len(self.timestamps) == len(
                 self.data
-            ), "Length of timestamps do not match the data"
+            ), f"Length of timestamps {len(self.timestamps)} do not match the data {len(self.data)}"
             if any(
                 self.timestamps[i] >= self.timestamps[i + 1]
                 for i in range(len(self.timestamps) - 1)
