### 2021-07-05 36008a3
better default timestamp handling
