### 2021-07-05 9350611
added test and check for empty dataset writing
