### 2021-07-05 ddedc5e
added test for writing dataset with empty frame
