+++ tests/unit/test_file_io_dataset.py
@@ -2,8 +2,10 @@ from pathlib import Path
 
 import pytest
 import pytest_check as check
+import pandas as pd
+import numpy as np
 
-from pointcloudset import Dataset
+from pointcloudset import Dataset, PointCloud
 
 
 def test_from_bag_wrong_topic(testbag1):
@@ -37,3 +39,28 @@ def test_empty_dataset(tmp_path: Path):
     testfile_name = tmp_path.joinpath("dataset0")
     with pytest.raises(ValueError):
         complete_empty_dataset.to_file(file_path=testfile_name, use_orig_filename=False)
+
+
+def test_dataset_with_empty_frame(testpointcloud_mini_real: PointCloud, tmp_path: Path):
+    fake_empty_df = pd.DataFrame.from_dict(
+        {
+            "x": [np.nan],
+            "y": [np.nan],
+            "z": [np.nan],
+            "intensity": [np.nan],
+            "t": [np.nan],
+            "reflectivity": [np.nan],
+            "ring": [np.nan],
+            "noise": [np.nan],
+            "range": [np.nan],
+            "original_id": [np.nan],
+        }
+    )
+    pc_empty = PointCloud(data=fake_empty_df)
+    testfile_name = tmp_path.joinpath("dataset")
+    ds = Dataset.from_instance("pointclouds", [testpointcloud_mini_real, pc_empty])
+    ds.to_file(file_path=testfile_name, use_orig_filename=False)
+    check.equal(testfile_name.exists(), True)
+    read_dataset = Dataset.from_file(testfile_name)
+    check.is_instance(read_dataset, Dataset)
+    check.equal(len(ds), len(read_dataset))
