### 2021-07-05 5b02797
added empty pointcloud object
