+++ tests/unit/test_pointcloud.py
@@ -16,6 +16,13 @@ def test_init(testpointcloud_mini_df: pd.DataFrame):
     check.equal(type(pointcloud), PointCloud)
 
 
+def test_empty_pointcloud():
+    empty_pc = PointCloud()
+    check.equal(type(empty_pc), PointCloud)
+    check.is_false(empty_pc._has_data())
+    check.equal(len(empty_pc), 0)
+
+
 def test_has_data(testpointcloud_mini: PointCloud):
     check.equal(testpointcloud_mini._has_data(), True)
 
@@ -51,6 +58,7 @@ def test_points2(testpointcloud_mini):
 
 def test_timestamp(testpointcloud_mini):
     check.equal(type(testpointcloud_mini.timestamp), datetime)
+    check.equal(testpointcloud_mini.timestamp, datetime(2020, 1, 1))
 
 
 def test_timestamp2(testpointcloud_mini):
@@ -60,7 +68,7 @@ def test_timestamp2(testpointcloud_mini):
     check.less(pc0.timestamp, pc1.timestamp)
 
 
-def test_org_file(testpointcloud):
+def test_org_file(testpointcloud, testset):
     check.equal(type(testpointcloud.orig_file), str)
     check.equal(Path(testpointcloud.orig_file).stem, "test")
 
