### 2021-07-05 6aa6532
added empty pointcloud object
