### 2021-07-06 5226c22
added test for apply which results in empty frame
