### 2021-07-07 cab9d29
added functionality to save dataset with empty frames
