### 2021-07-07 45e017a
added functionality to save dataset with empty frames
