+++ doc/sphinx/source/conf.py
@@ -6,14 +6,13 @@
 
 # -- Path setup --------------------------------------------------------------
 
+import codecs
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #
 import os
 import sys
-import codecs
-
 
 sys.path.insert(0, os.path.abspath("../../../pointcloudset"))
 
