+++ pointcloudset/dataset.py
@@ -76,11 +76,13 @@ class Dataset(DatasetCore):
             ValueError: If file format is not supported.
             TypeError: If file_path is not a Path object.
         """
+        from_dir = False
         if not isinstance(file_path, Path):
             raise TypeError("Expecting a Path object for file_path")
         ext = file_path.suffix[1:].upper()
         if ext == "":
             ext = "DIR"
+            from_dir = True
         if ext not in DATASET_FROM_FILE:
             raise ValueError(
                 (
@@ -89,8 +91,11 @@ class Dataset(DatasetCore):
                 )
             )
         res = DATASET_FROM_FILE[ext](file_path, **kwargs)
-        out = cls(data=res["data"], timestamps=res["timestamps"], meta=res["meta"])
-        return out._replace_nan_frames_with_empty()
+        meta = res["meta"]
+        out = cls(data=res["data"], timestamps=res["timestamps"], meta=meta)
+        if from_dir:
+            out = out._replace_nan_frames_with_empty(res["empty_data"])
+        return out
 
     def to_file(self, file_path: Path = Path(), **kwargs) -> None:
         """Writes a Dataset to a file.
@@ -467,27 +472,25 @@ class Dataset(DatasetCore):
         self._check()
         return self
 
-    def _replace_empty_frames_with_nan(self):
+    def _replace_empty_frames_with_nan(self, empty_data: pandas.DataFrame):
         """Function to replace empty pointclouds with pointclouds wiht 1 point with all
         nan values. Needed to save files with dask.
         """
 
-        default_data = pandas.DataFrame(self[0].data.iloc[0]).T
-
         def _exchange_empty_pointclouds_with_nan(frame: PointCloud) -> PointCloud:
             if not frame._has_data():
-                frame.data = default_data
+                frame.data = empty_data
             return frame
 
         return self.apply(_exchange_empty_pointclouds_with_nan)
 
-    def _replace_nan_frames_with_empty(self):
+    def _replace_nan_frames_with_empty(self, empty_data: pandas.DataFrame):
         """Function to replace nan pointclouds with empty pointcouds
         Needed to after reading dataset files.
         """
 
         def _exchange_nan_pointclouds_with_empty(frame: PointCloud) -> PointCloud:
-            if (len(frame) == 1) and frame.data.isnull().all().all():
+            if (len(frame) == 1) and np.allclose(empty_data.values, frame.data.values):
                 frame = PointCloud(columns=frame.data.columns)
             return frame
 
