+++ tests/unit/test_dataset.py
@@ -454,17 +454,34 @@ def test_dataset_mean(
     )
 
 
+empty_data = pd.DataFrame.from_dict(
+    {
+        "x": {0: 0.0},
+        "y": {0: -0.0},
+        "z": {0: 0.0},
+        "intensity": {0: 0.0},
+        "t": {0: 0.0},
+        "reflectivity": {0: 0.0},
+        "ring": {0: 0.0},
+        "noise": {0: 74.0},
+        "range": {0: 0.0},
+    }
+)
+
+
 def test_replace_empty_frames_with_nan(testdataset_with_empty_frame: Dataset):
-    test = testdataset_with_empty_frame._replace_empty_frames_with_nan()
+    test = testdataset_with_empty_frame._replace_empty_frames_with_nan(
+        empty_data=empty_data
+    )
     check.equal(len(test), len(testdataset_with_empty_frame))
     check.equal(len(test[1]), 1)
 
 
 def test_replace_nan_frames_with_empty(testdataset_with_empty_frame: Dataset):
-    test = testdataset_with_empty_frame._replace_empty_frames_with_nan()
-    test2 = test._replace_nan_frames_with_empty()
+    test = testdataset_with_empty_frame._replace_empty_frames_with_nan(
+        empty_data=empty_data
+    )
+    test2 = test._replace_nan_frames_with_empty(empty_data=empty_data)
     check.equal(len(test2), len(testdataset_with_empty_frame))
     check.is_false(test2[1]._has_data())
     check.equal(len(test2[1]), 0)
-
-
