### 2021-07-08 9ce26a0
backwards compatibility of dir files and version push
