+++ pointcloudset/__init__.py
@@ -1,4 +1,4 @@
 from .dataset import Dataset
 from .pointcloud import PointCloud
 
-__version__ = "0.2.2"
+__version__ = "0.2.3"
