### 2021-07-10 #3471 @whedon
[REVIEW]: pointcloudset: Efficient analysis of large datasets of point clouds recorded over time
**Submitting author:** @tgoelles (<a href="http://orcid.org/0000-0002-3925-6260">Thomas Goelles</a>)
**Repository:** <a href="https://github.com/virtual-vehicle/pointcloudset/" target ="_blank">https://github.com/virtual-vehicle/pointcloudset/</a>
**Version:** v0.3.3

- [x] #3570

- [x] #3571

- [x] #3569

- [x] #3573
- [x] #3572
