### 2021-07-14 virtual-vehicle/pointcloudset#8 @RonaldEnsing
FileNotFoundError in Working_with_kitti_dataset tutorial
I am writing this issue as a JOSS reviewer. See more info here: https://github.com/openjournals/joss-reviews/issues/3471.

The 'Working_with_kitti_dataset' gives the following error:
`FileNotFoundError: [Errno 2] No such file or directory: '/pointcloudset/tutorial_notebooks/kitti_2011_09_26_drive_0002_synced.bag'`

I am using the pointcloudset docker container. Since the python code downloads the data, I'd expect this to work 'out-of-the-box', without the need to configure the paths manually. The bag file seems to be missing entirely.
