### 2021-07-14 virtual-vehicle/pointcloudset#8 @tgoelles
Hi. Thanks for reviewing our package!

Yes we didn't want to include the bag file in the repo as it is too large and to avoid any license issues. Also KITTI does not provide bag files directly. In order to get the data you need to run cell 3 in [Working_with_kitti_dataset.ipynb](https://github.com/virtual-vehicle/pointcloudset/blob/master/doc/sphinx/source/tutorial_notebooks/Working_with_kitti_dataset.ipynb)

With  uncommented `docker run ...` line like this:

```
!wget https://s3.eu-central-1.amazonaws.com/avg-kitti/raw_data/2011_09_26_drive_0002/2011_09_26_drive_0002_sync.zip
!wget https://s3.eu-central-1.amazonaws.com/avg-kitti/raw_data/2011_09_26_calib.zip
!unzip 2011_09_26_drive_0002_sync.zip
!unzip 2011_09_26_calib.zip
!docker run -v `pwd`:/data -it yoyo860224/kitti2bag -t 2011_09_26 -r 0002 raw_synced
```

This will download the data and convert it to a bagfile with the help of yoyo860224/kitti2bag.
The other notebooks should work out of the box, as they are also included in the automatic tests.

(@tobiashammer)
