### 2021-07-14 virtual-vehicle/pointcloudset#8 @RonaldEnsing
Ah, I see. What is the reason for commenting the `kitti2bag` docker command?

I am running the notebook in a docker container. So the `docker run ...` would become a nested docker command.

Would it be an idea to make `kitti2bag` available in the pointcloudset docker image so that the tutorials can be run directly from the pointcloudset containers?
