### 2021-07-14 virtual-vehicle/pointcloudset#8 @tgoelles
Yes thats the reason it is commented, to avoid issues with nested docker commands. So you could run the whole command of cell 3 outside of docker to get the bag file.

kitti2bag is quite hard to install and we want to avoid the additional complexity. We included the notebook just to show that pointcloudset also works with KITTI if needed.
