### 2021-07-14 virtual-vehicle/pointcloudset#9 @RonaldEnsing
Suggestions for the README
Related to https://github.com/openjournals/joss-reviews/issues/3471.

I'd recommend to add links to refer to some things mentioned in the readme. For example:
- A link to the pypi website of pointcloudset: https://pypi.org/project/pointcloudset/.
- A link to the HTML documentation: https://virtual-vehicle.github.io/pointcloudset/.

No default tag `latest` is defined for the https://hub.docker.com/repository/docker/tgoelles/pointcloudset image, so a `docker pull tgoelles/pointcloudset` will not work. A tag would need to be specified by the user. Is this done intentionally? In that case I think that it would be good to write the exact docker pull command in the readme.

The Quickstart example shows a code snippet that refers to `rosbag_file.bag` and `lasfile.las`. Are these files available somewhere? Would it be possible to include a quickstart example with existing example files - and ideally - some example output that includes the visualizations as shown in the README?
