### 2021-07-14 virtual-vehicle/pointcloudset#8 @RonaldEnsing
OK, clear. Maybe it would be a good idea to clarify this in the notebook?
