### 2021-07-14 #3471 @RonaldEnsing
I have added a few suggestions to the README. My review is completed. I think this is a great tool and very well documented!
