### 2021-07-14 virtual-vehicle/pointcloudset#9 @tgoelles
Yes links are certainly a good idea.

Yes I avoided the latest tag since it can cause some confusion. But, I think I can easily add it to github actions to tag it both with the version name and latest.

The quickstart example is only there to give a quick idea what the API looks like. For working examples of code please see [usage.ipynb](https://github.com/virtual-vehicle/pointcloudset/blob/master/doc/sphinx/source/tutorial_notebooks/usage.ipynb). Do you think that the user expects working examples in the README or just a general idea of what the API looks like?

I also want to change the sphinx setup to run the notebooks so that notebook is populated with results and plots. This would also improve the html documentation.
