### 2021-07-14 virtual-vehicle/pointcloudset#9 @RonaldEnsing
> Yes I avoided the latest tag since it can cause some confusion. But, I think I can easily add it to github actions to tag it both with the version name and latest.

I think that would be nice. I expect that most users will do a `docker pull` without specifying or looking into versions and expect it to work. Especially when they are only exploring the package to see if it is useful to them.

> Do you think that the user expects working examples in the README or just a general idea of what the API looks like?

I'd expect a working example. If the quickstart is only meant to show a general idea, then what general idea do you want to show specifically? It only shows the file I/O. I think it would be nice if a quickstart allows someone to reproduce the Figures under 'Features' in the README. If that would make the README too cluttered, then maybe a link to a notebook on how to produce these Figures?

> I also want to change the sphinx setup to run the notebooks so that notebook is populated with results and plots. This would also improve the html documentation.

Sounds good! I think the notebooks are somewhat 'hidden' in the docs directory. Maybe it would help to add a link from the README to the notebooks to increase the discoverability of the notebooks.
