### 2021-07-14 virtual-vehicle/pointcloudset#8 @tobiashammer
Hi, yes I will update the notebook.
