### 2021-07-15 0fd8a49
update kitti notebook (issue #8)
