### 2021-07-15 d5dd26f
run notebooks in doc
