### 2021-07-15 955c4e9
added latest tag for pointcloudset image
