### 2021-07-15 0972855
added working example in readme and include it in tests
