### 2021-07-15 37eef95
added working example in readme and include it in tests
