### 2021-07-15 ca5212f
changed quick start and gif
