### 2021-07-15 73ea474
fixed bug with plot hover_data=True
