### 2021-07-15 virtual-vehicle/pointcloudset#9 @tgoelles
We fixed everything except a problem with updating documentation from notebooks remains. See issue #12
