### 2021-08-05 #3471 @danielskatz
@hechth - you will need to manually open these issues in the software repository.  The GitHub feature you used opens them in the JOSS reviews repository, which is only used for the overall review of all JOSS papers.
