### 2021-08-05 #3471 @danielskatz
When you opened the issues, I thought you used the GitHub "convert to issue" feature, which doesn't work with JOSS, as it opens the issues in the wrong repository (the one where JOSS reviews go, and where automated JOSS processing works on them), rather than the source repository (in this case, https://github.com/virtual-vehicle/pointcloudset/), where issues need to opened manually (via https://github.com/virtual-vehicle/pointcloudset/issues/new/choose), so I closed all the issues in the wrong repository.  However, it now appears you used both methods, so if the issues in are https://github.com/virtual-vehicle/pointcloudset/issues correct, then everything is ok.
