### 2021-08-06 d3404f3
Update ci-environment.yml
added open3d to environment
