+++ doc/sphinx/ci-environment.yml
@@ -5,6 +5,7 @@ channels:
   - open3d-admin
 dependencies:
   - numpy
+  - open3d=0.12.0
   - pandas
   - pip
   - python>=3.8
