+++ .github/workflows/publish_on_conda.yml
@@ -1,4 +1,4 @@
-name: publish_conda
+name: publish_on_conda
 
 on:
   push:
