### 2021-08-09 81b69fe
test conda build with custom channels
