+++ conda/conda_build_config.yaml
@@ -1,2 +1,7 @@
+channels:
+  - defaults
+  - conda-forge
+  - open3d-admin
+
 python:
   - 3.8
