### 2021-08-09 a108bfe
changed conde ci version to v1.3_open3d
