### 2021-08-09 78569d4
delte channel specs in conda_build_config
