+++ conda/conda_build_config.yaml
@@ -1,7 +1,2 @@
-channels:
-  - defaults
-  - conda-forge
-  - open3d-admin
-
 python:
   - 3.8
