### 2021-08-10 006eca3
refactored test_diff
