### 2021-08-11 251c3e7
refactore test_geometry_plane
