+++ tests/unit/test_geometry_plane.py
@@ -6,37 +6,24 @@ from numpy.testing import assert_array_equal
 from pointcloudset.geometry import plane
 
 
-def test_distance_to_point1():
+@pytest.mark.parametrize(
+    "point, res", [([0, 0, 0], 10), ([1, 1, 1], 11), ([-1, -1, -1], 9)]
+)
+def test_distance_to_point(point, res):
     distance = plane.distance_to_point(
-        point_A=np.array([0, 0, 0]), plane_model=np.array([1, 0, 0, 10])
+        point_A=np.array(point), plane_model=np.array([1, 0, 0, 10])
     )
-    check.equal(distance, 10)
-    distance = plane.distance_to_point(
-        point_A=np.array([1, 1, 1]), plane_model=np.array([1, 0, 0, 10])
-    )
-    check.equal(distance, 11)
-    distance = plane.distance_to_point(
-        point_A=np.array([-1, -1, -1]), plane_model=np.array([1, 0, 0, 10])
-    )
-    check.equal(distance, 9)
+    check.equal(distance, res)
 
 
-def test_distance_to_point2():
+@pytest.mark.parametrize("normal_dist, res", [(False, np.sqrt(2)), (True, 1.0)])
+def test_distance_to_point_normal_dist(normal_dist, res):
     distance = plane.distance_to_point(
         point_A=np.array([2, 2, 0]),
         plane_model=np.array([1, 0, 0, -1]),
-        normal_dist=False,
+        normal_dist=normal_dist,
     )
-    check.equal(distance, (np.sqrt(2)))
-
-
-def test_distance_to_point3():
-    distance = plane.distance_to_point(
-        point_A=np.array([2, 2, 0]),
-        plane_model=np.array([1, 0, 0, -1]),
-        normal_dist=True,
-    )
-    check.equal(distance, 1.0)
+    check.equal(distance, res)
 
 
 def test_distance_to_point_error1():
