### 2021-08-11 9ebf812
refactor test_geometry_point
