### 2021-08-11 a447282
refactor test_plane_segmentation
