### 2021-08-11 2f1c075
refactor test_pointcloud
