### 2021-08-12 fb73e37
added limit to plot pointclouds
