### 2021-08-12 bfe1b0e
added random_down_sample and suggest it for large plots
