### 2021-08-12 795442f
added random_down_sample and suggest it for large plots
