+++ conda/environment.yml
@@ -1,4 +1,4 @@
-name: base
+name: pointcloudset
 channels:
   - defaults
   - conda-forge
