+++ .github/workflows/tests.yml
@@ -20,7 +20,7 @@ jobs:
             hashFiles('etc/example-environment.yml') }}
       - uses: conda-incubator/setup-miniconda@v2
         with:
-          activate-environment: test
+          activate-environment: pointcloudset
           auto-activate-base: false
           environment-file: conda/environment.yml
           use-only-tar-bz2: true # IMPORTANT: This needs to be set for caching to work properly!
