### 2021-08-16 710b27f
added version numbers to the base docker container
