### 2021-08-17 be4c706
docker buid with custom solution
