+++ .github/workflows/docker.yml
@@ -1,54 +1,43 @@
-name: Docker base image
+name: Docker base
 
 on:
-  push:
-    branches:
-      - main
+  release:
+    types: # This configuration does not affect the page_build event above
+      - created
+
+
+  # Allows you to run this workflow manually from the Actions tab
+  workflow_dispatch:
+    inputs:
+      name:
+        description: "docker tag"
+        required: true
+        default: "latest"
+
+env:
+  image_name: pointcloudset_base
+
 jobs:
-
   build:
-
     runs-on: ubuntu-latest
 
     steps:
       - uses: actions/checkout@v2
-        with:
-            fetch-depth: 0  # OR "2" -> To retrieve the preceding commit.
-      - name: Get changed files using defaults
-        id: changed-files
-        uses: tj-actions/changed-files@v6.3
-
-      - name: Get changed files using a comma separator
-        id: changed-files-comma
-        uses: tj-actions/changed-files@v6.3
-        with:
-          separator: ","
-
-      - name: List all added files
-        run: |
-          for file in "${{ steps.changed-files.outputs.added_files }}"; do
-            echo "$file was added"
-          done
-
-      - name: Get specific changed files
-        id: changed-files-specific
-        uses: tj-actions/changed-files@v6.3
-        with:
-          files: |
-            docker/Dockerfile
-            conda/environment.yml
-
+      - name: Get the tag name
+        if: github.event_name != 'workflow_dispatch'
+        run: echo "TAG=${GITHUB_REF/refs\/tags\//}" >> $GITHUB_ENV
+      - name: manual tag
+        if: github.event_name == 'workflow_dispatch'
+        run: echo "TAG=${{ github.event.inputs.name }}" >> $GITHUB_ENV
       - name: Login to DockerHub
         uses: docker/login-action@v1
         with:
           username: ${{ secrets.DOCKERHUB_USERNAME }}
           password: ${{ secrets.DOCKERHUB_TOKEN }}
-
-      - name: Build and push tgoelles/pointcloudset_base
-        if: steps.changed-files-specific.outputs.any_changed == 'true'
+      - name: Build and push tgoelles/pointcloudset
         run: |
-          docker build . --file docker/release.Dockerfile --tag tgoelles/pointcloudset_base:${{ env.TAG }} --tag tgoelles/pointcloudset_base:latest
-          docker push tgoelles/pointcloudset_base:${{ env.TAG }}
-          docker push tgoelles/pointcloudset_base:latest
-
-
+          docker build . --file docker/release.Dockerfile --tag tgoelles/${{ env.IMAGE_NAME }}:${{ env.TAG }} --tag tgoelles/${{ env.IMAGE_NAME }}:latest
+          docker push tgoelles/${{ env.IMAGE_NAME }}:${{ env.TAG }}
+      - name: push latest tag
+        if: github.event_name != 'workflow_dispatch'
+        run:  docker push tgoelles/pointcloudset:latest
