### 2021-08-17 0fb7a7b
docker buid with custom solution
