+++ .github/workflows/docker_release.yml
@@ -1,4 +1,4 @@
-name: Docker base
+name: Docker release
 
 on:
   release:
@@ -9,35 +9,43 @@ on:
   # Allows you to run this workflow manually from the Actions tab
   workflow_dispatch:
     inputs:
-      name:
+      tag:
         description: "docker tag"
         required: true
-        default: "latest"
-
-env:
-  image_name: pointcloudset_base
+        default: "test"
 
 jobs:
-  build:
+  docker:
     runs-on: ubuntu-latest
-
     steps:
-      - uses: actions/checkout@v2
+      -
+        name: Checkout
+        uses: actions/checkout@v2
       - name: Get the tag name
         if: github.event_name != 'workflow_dispatch'
         run: echo "TAG=${GITHUB_REF/refs\/tags\//}" >> $GITHUB_ENV
       - name: manual tag
         if: github.event_name == 'workflow_dispatch'
-        run: echo "TAG=${{ github.event.inputs.name }}" >> $GITHUB_ENV
-      - name: Login to DockerHub
+        run: echo "TAG=${{ github.event.inputs.tag }}" >> $GITHUB_ENV
+      -
+        name: Set up QEMU
+        uses: docker/setup-qemu-action@v1
+      -
+        name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v1
+      -
+        name: Login to DockerHub
         uses: docker/login-action@v1
         with:
           username: ${{ secrets.DOCKERHUB_USERNAME }}
           password: ${{ secrets.DOCKERHUB_TOKEN }}
-      - name: Build and push tgoelles/pointcloudset
-        run: |
-          docker build . --file docker/release.Dockerfile --tag tgoelles/${{ env.IMAGE_NAME }}:${{ env.TAG }} --tag tgoelles/${{ env.IMAGE_NAME }}:latest
-          docker push tgoelles/${{ env.IMAGE_NAME }}:${{ env.TAG }}
-      - name: push latest tag
-        if: github.event_name != 'workflow_dispatch'
-        run:  docker push tgoelles/pointcloudset:latest
+      -
+        name: Build and push
+        uses: docker/build-push-action@v2
+        with:
+          context: .
+          push: true
+          platforms: linux/amd64,linux/arm64
+          tags:  |
+              tgoelles/pointcloudset:latest
+              tgoelles/pointcloudset:${{ env.TAG }}
