+++ .github/workflows/docker_base.yml
@@ -44,7 +44,7 @@ jobs:
         uses: docker/build-push-action@v2
         with:
           context: .
-          push: true
+          push: false
           platforms: linux/amd64,linux/arm64
           tags:  |
               tgoelles/pointcloudset_base:latest
