+++ .github/workflows/docker_base.yml
@@ -43,7 +43,7 @@ jobs:
         name: Build and push
         uses: docker/build-push-action@v2
         with:
-          context: .
+          context: docker
           push: false
           platforms: linux/amd64,linux/arm64
           tags:  |
