### 2021-08-17 06c5fb2
changed paths back to . for docker_base
