### 2021-08-17 ea9fb74
changed paths back to . for docker_base
