+++ paper/paper.bib
@@ -125,3 +125,30 @@
   doi = {10.25080/majora-7b98e3ed-013}
 }
 
+
+@misc{pdal_contributors_2018_2556738,
+author = {{PDAL Contributors}},
+title = {PDAL Point Data Abstraction Library},
+month = nov,
+year = 2018,
+doi = {10.5281/zenodo.2556738},
+url = {https://doi.org/10.5281/zenodo.2556738}
+}
+
+@InProceedings{Rusu_ICRA2011_PCL,
+  author    = {Radu Bogdan Rusu and Steve Cousins},
+  title     = {{3D is here: Point Cloud Library (PCL)}},
+  booktitle = {{IEEE International Conference on Robotics and Automation (ICRA)}},
+  month     = {May 9-13},
+  year      = {2011},
+  address   = {Shanghai, China},
+  publisher = {IEEE}
+}
+
+@software{python-pcl,
+author = {{Python-PCL Contributors}},
+title = {Python-PCL: Python bindings to the pointcloud library (pcl)},
+url = {https://github.com/strawlab/python-pcl},
+version = {v0.2.0},
+date = {2017-12-07},
+}
