+++ paper/paper.md
@@ -41,7 +41,9 @@ Considering recently emerging, promising lidar technologies, such as micro-elect
 
 Apart from the progress in the lidar sector, technological improvements, as well as size and cost reduction can also be observed for other 3D sensing technologies, such as radar and RGB-D cameras. This will additionally open up new possibilities and application areas for 3D sensing methods and increases the importance of python packages that are able to process large amounts of point cloud data.
 
-Other python packages for point clouds, such as open3D and pyntcloud, focus on processing single point clouds. ROS (robot operating system) [@ros:2018] provides a way to store, access, and visualize multiple point clouds stored as `rosbags`. However, these `rosbags` are meant to be accessed only in a serial fashion, which is not ideal for post processing and not well suited for extracting subsets of the point cloud dataset.
+Other python packages for point clouds, such as open3D, pyntcloud, and pcl [@Rusu_ICRA2011_PCL] and its python bindings [@python-pcl], focus on processing single point clouds rather than on processing time series of point clouds. Another library is PDAL [@pdal_contributors_2018_2556738] which also works with pipelines on point clouds. However, it is focused on single point cloud processing as well. ROS (robot operating system) [@ros:2018] provides a way to store, access, and visualize multiple point clouds stored as `rosbags`. However, these `rosbags` are meant to be accessed only in a serial fashion, which is not ideal for post processing and not well suited for extracting subsets of the point cloud dataset.
+
+Compared to mentioned packages, pointcloudset provides efficient analysis of time series of point clouds by parallel processing. E.g. the package is a helpful toolkit for post processing of lidar datasets recorded by ROS or for post processing of multiple lidar scans from terrestrial laser scanners. Algorithms can be developed for a single point cloud and can then be applied to big datasets of point clouds. 
 
 
 ![Dataset object with main properties and ways to read and write data. \label{fig:dataset}](./figures/data_pipeline2.pdf){ width=100% }
