+++ .github/workflows/docker_base.yml
@@ -45,7 +45,7 @@ jobs:
         with:
           context: .
           file: docker/Dockerfile
-          push: false
+          push: true
           platforms: linux/amd64
           tags:  |
               tgoelles/pointcloudset_base:latest
