+++ .github/workflows/docker_release.yml
@@ -44,8 +44,9 @@ jobs:
         uses: docker/build-push-action@v2
         with:
           context: .
-          push: true
-          platforms: linux/amd64,linux/arm64
+          file: docker/release.Dockerfile
+          push: false
+          platforms: linux/amd64
           tags:  |
               tgoelles/pointcloudset:latest
               tgoelles/pointcloudset:${{ env.TAG }}
