+++ conda/meta.yaml
@@ -1,5 +1,5 @@
 {% set name = "pointcloudset" %}
-{% set version = "0.2.3" %}
+{% set version = "0.3.0" %}
 
 package:
   name: "{{ name|lower }}"
