### 2021-08-17 44ab972
merged all docker actions into one
