+++ .github/workflows/docker.yml
@@ -39,8 +39,20 @@ jobs:
         with:
           username: ${{ secrets.DOCKERHUB_USERNAME }}
           password: ${{ secrets.DOCKERHUB_TOKEN }}
+
       -
-        name: Build and push
+        name: Build and push the base image
+        uses: docker/build-push-action@v2
+        with:
+          context: .
+          file: docker/Dockerfile
+          push: true
+          platforms: linux/amd64
+          tags:  |
+              tgoelles/pointcloudset_base:latest
+              tgoelles/pointcloudset_base:${{ env.TAG }}
+      -
+        name: Build and push the release image
         uses: docker/build-push-action@v2
         with:
           context: .
