### 2021-08-17 bb42ddc
merged all docker actions into one
