+++ .github/workflows/docker_base.yml
@@ -1,52 +1,0 @@
-name: Docker base
-
-on:
-  release:
-    types: # This configuration does not affect the page_build event above
-      - created
-
-
-  # Allows you to run this workflow manually from the Actions tab
-  workflow_dispatch:
-    inputs:
-      tag:
-        description: "docker tag"
-        required: true
-        default: "test"
-
-jobs:
-  docker:
-    runs-on: ubuntu-latest
-    steps:
-      -
-        name: Checkout
-        uses: actions/checkout@v2
-      - name: Get the tag name
-        if: github.event_name != 'workflow_dispatch'
-        run: echo "TAG=${GITHUB_REF/refs\/tags\//}" >> $GITHUB_ENV
-      - name: manual tag
-        if: github.event_name == 'workflow_dispatch'
-        run: echo "TAG=${{ github.event.inputs.tag }}" >> $GITHUB_ENV
-      -
-        name: Set up QEMU
-        uses: docker/setup-qemu-action@v1
-      -
-        name: Set up Docker Buildx
-        uses: docker/setup-buildx-action@v1
-      -
-        name: Login to DockerHub
-        uses: docker/login-action@v1
-        with:
-          username: ${{ secrets.DOCKERHUB_USERNAME }}
-          password: ${{ secrets.DOCKERHUB_TOKEN }}
-      -
-        name: Build and push
-        uses: docker/build-push-action@v2
-        with:
-          context: .
-          file: docker/Dockerfile
-          push: true
-          platforms: linux/amd64
-          tags:  |
-              tgoelles/pointcloudset_base:latest
-              tgoelles/pointcloudset_base:${{ env.TAG }}
