### 2021-08-17 2316578
generate pointcloudset environment
