+++ CHANGELOG.rst
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_,
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
+
+[0.3.1] - 2021-08-17
+-------------
+
+Fixed
+~~~~~~
+- conda environment name was still "base" now is "pointcloudset"
+
+
 [0.3.0] - 2021-08-17
 -------------
 
