+++ CHANGELOG.rst
@@ -6,7 +6,9 @@ The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
 
-[0.3.1] - 2021-08-17
+
+
+[Unreleased]
 -------------
 
 Fixed
