+++ .bumpversion.cfg
@@ -1,0 +1,15 @@
+[bumpversion]
+current_version = 0.3.0
+commit = False
+tag = True
+
+[bumpversion:file:pointcloudset/__init__.py]
+
+[bumpversion:file:conda/meta.yaml]
+search = set version = "{current_version}"
+replace = set version = "{new_version}"
+
+
+[bumpversion:file:CHANGELOG.rst]
+search = %%version%% (unreleased)
+replace = v{new_version} ({now:%Y-%m-%d})
