+++ .bumpversion.cfg
@@ -11,5 +11,5 @@ replace = set version = "{new_version}"
 
 
 [bumpversion:file:CHANGELOG.rst]
-search = %%version%% (unreleased)
-replace = v{new_version} ({now:%Y-%m-%d})
+search = Unreleased
+replace = {new_version} ({now:%Y-%m-%d})
