+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.3.0
+current_version = 0.3.1
 commit = False
 tag = True
 
@@ -9,7 +9,6 @@ tag = True
 search = set version = "{current_version}"
 replace = set version = "{new_version}"
 
-
 [bumpversion:file:CHANGELOG.rst]
 search = Unreleased
-replace = {new_version} ({now:%Y-%m-%d})
+replace = {new_version} - ({now:%Y-%m-%d})
