+++ .bumpversion.cfg
@@ -1,6 +1,6 @@
 [bumpversion]
-current_version = 0.3.1
-commit = False
+current_version = 0.3.0
+commit = True
 tag = True
 
 [bumpversion:file:pointcloudset/__init__.py]
