+++ .github/workflows/docker.yml
@@ -1,9 +1,11 @@
 name: Docker release
 
 on:
-  release:
-    types: # This configuration does not affect the page_build event above
-      - created
+  push:
+    branches:
+      - "!*"
+    tags:
+      - "v*"
 
 
   # Allows you to run this workflow manually from the Actions tab
