### 2021-08-17 #3471 @hechth
@tgoelles @birgitschlager Thank you very much for the additions, it seems a lot clearer to me now!

I will investigate the functionality and performance a bit more during this week but I think it will be fine - let's wait if maybe the dependencies will be put into conda so that the package can be published there. If not, this is definitely not an issue that would block the acceptance!
