### 2021-08-18 7206b18
use explicit version number in Dockerfile for development
