+++ .bumpversion.cfg
@@ -5,6 +5,10 @@ tag = True
 
 [bumpversion:file:pointcloudset/__init__.py]
 
+[bumpversion:file:.devcontainer/Dockerfile]
+search = tgoelles/pointcloudset_base:v"{current_version}"
+replace = tgoelles/pointcloudset_base:v"{new_version}"
+
 [bumpversion:file:conda/meta.yaml]
 search = set version = "{current_version}"
 replace = set version = "{new_version}"
