### 2021-08-18 f7863fe
better handling of pointcloudset conda env
