+++ docker/Dockerfile
@@ -81,10 +81,12 @@ COPY conda/environment.yml* /tmp/conda-tmp/
 RUN /opt/conda/bin/conda update -n base -c defaults conda
 #RUN /opt/conda/bin/conda config --set channel_priority strict
 # Update Python environment based on environment.yml (if presenet)
-RUN if [ -f "/tmp/conda-tmp/environment.yml" ]; then /opt/conda/bin/conda env create -f /tmp/conda-tmp/environment.yml; fi \
-    && rm -rf /tmp/conda-tmp \
-    # install the kernel to use with jupyter and some extensions
-    && python -m ipykernel install --user --name pointcloudset
+RUN /opt/conda/bin/conda env create -f /tmp/conda-tmp/environment.yml \
+    && rm -rf /tmp/conda-tmp
+RUN echo "conda activate pointcloudset" >> ~/.bashrc
+SHELL ["/bin/bash", "--login", "-c"]
+# install the kernel to use with jupyter and some extensions
+RUN python -m ipykernel install --user --name pointcloudset
 
 # ROSPY loggging configuation file
 COPY docker/python_logging.conf /etc/ros/
