### 2021-08-18 edf884d
current test to activate pointcloudset env
