### 2021-08-18 bfb08b5
current test with activate pointcloudset
