+++ docker/Dockerfile
@@ -92,3 +92,5 @@ RUN /bin/bash -c  'source activate pointcloudset && \
 COPY docker/python_logging.conf /etc/ros/
 # Switch back to dialog for any ad-hoc use of apt-get
 ENV DEBIAN_FRONTEND=noninteractive
+
+ENTRYPOINT ["conda", "run", "-n", "pointcloudset"]
