### 2021-08-18 6f8a59e
activate pointcloudset automatically
