+++ .bumpversion.cfg
@@ -5,7 +5,7 @@ tag = True
 
 [bumpversion:file:pointcloudset/__init__.py]
 
-[bumpversion:file:.devcontainer/Dockerfile]
+[bumpversion:file:.devcontainer/dev.Dockerfile]
 search = tgoelles/pointcloudset_base:v"{current_version}"
 replace = tgoelles/pointcloudset_base:v"{new_version}"
 
