### 2021-08-18 d68bd8b
fixed bumpversion setup for dev.Dockerfile
