+++ .bumpversion.cfg
@@ -6,8 +6,8 @@ tag = True
 [bumpversion:file:pointcloudset/__init__.py]
 
 [bumpversion:file:.devcontainer/dev.Dockerfile]
-search = tgoelles/pointcloudset_base:v"{current_version}"
-replace = tgoelles/pointcloudset_base:v"{new_version}"
+search = tgoelles/pointcloudset_base:v{current_version}
+replace = tgoelles/pointcloudset_base:v{new_version}
 
 [bumpversion:file:conda/meta.yaml]
 search = set version = "{current_version}"
