+++ conda/meta.yaml
@@ -1,5 +1,5 @@
 {% set name = "pointcloudset" %}
-{% set version = "0.3.0" %}
+{% set version = "0.3.1" %}
 
 package:
   name: "{{ name|lower }}"
