+++ CHANGELOG.rst
@@ -8,13 +8,17 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 
 
 
-0.3.1 - (2021-08-18)
+Unreleased
 -------------
 
 Fixed
 ~~~~~~
 - conda environment name was still "base" now is "pointcloudset"
 - automatic start of pointcloudset conda environment now working
+
+Changed
+~~~~~~
+- use fixed version number of pointcloudset_base image
 
 
 0.3.0 (2021-08-17)
