+++ CHANGELOG.rst
@@ -8,7 +8,7 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 
 
 
-Unreleased
+0.3.1 - (2021-08-18)
 -------------
 
 Fixed
