### 2021-08-18 944024b
consistent dockerfile naming
