+++ .github/workflows/docker.yml
@@ -47,7 +47,7 @@ jobs:
         uses: docker/build-push-action@v2
         with:
           context: .
-          file: docker/Dockerfile
+          file: docker/base.Dockerfile
           push: true
           platforms: linux/amd64
           tags:  |
