+++ .bumpversion.cfg
@@ -1,11 +1,15 @@
 [bumpversion]
-current_version = 0.3.1
+current_version = 0.3.0
 commit = True
 tag = True
 
 [bumpversion:file:pointcloudset/__init__.py]
 
 [bumpversion:file:.devcontainer/dev.Dockerfile]
+search = tgoelles/pointcloudset_base:v{current_version}
+replace = tgoelles/pointcloudset_base:v{new_version}
+
+[bumpversion:file:.devcontainer/release.Dockerfile]
 search = tgoelles/pointcloudset_base:v{current_version}
 replace = tgoelles/pointcloudset_base:v{new_version}
 
