+++ .bumpversion.cfg
@@ -9,7 +9,7 @@ tag = True
 search = tgoelles/pointcloudset_base:v{current_version}
 replace = tgoelles/pointcloudset_base:v{new_version}
 
-[bumpversion:file:.devcontainer/release.Dockerfile]
+[bumpversion:file:docker/release.Dockerfile]
 search = tgoelles/pointcloudset_base:v{current_version}
 replace = tgoelles/pointcloudset_base:v{new_version}
 
