+++ .github/workflows/docker.yml
@@ -1,4 +1,4 @@
-name: Docker release
+name: Docker
 
 on:
   push:
