+++ conda/environment.yml
@@ -14,6 +14,8 @@ dependencies:
   - plotly
   - plotly-orca
   - ipykernel
+  - pdal
+  - fiona
   - pip:
       - fastparquet
       - jupyter
