### 2021-09-02 574d7f8
fixed point size issue with overlay plots
