+++ CHANGELOG.rst
@@ -6,7 +6,12 @@ The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
 
+Unreleased
+-------------
 
+Fixed
+~~~~~~
+- point_size option had no effect when using overlays
 
 0.3.2 - (2021-08-18)
 -------------
