### 2021-09-14 b66f4d6
Update tests.yml
setup of conda takes a long time. Testing solutios
