### 2021-09-14 5a6b3d0
Update tests.yml
setup mini conda failed
