### 2021-09-14 658c820
Update tests.yml
dont use strict priority
