+++ .github/workflows/tests.yml
@@ -22,7 +22,6 @@ jobs:
         with:
           activate-environment: pointcloudset
           auto-activate-base: false
-          channel-priority: strict
           environment-file: conda/environment.yml
           use-only-tar-bz2: true # IMPORTANT: This needs to be set for caching to work properly!
       - name: test conda
