### 2021-09-14 79117b2
Update tests.yml
test with actual hash file
