+++ .github/workflows/tests.yml
@@ -12,12 +12,12 @@ jobs:
       - name: Cache conda
         uses: actions/cache@v2
         env:
-          # Increase this value to reset cache if etc/example-environment.yml has not changed
+          # Increase this value to reset cache if yml has not changed
           CACHE_NUMBER: 1
         with:
           path: ~/conda_pkgs_dir
           key: ${{ runner.os }}-conda-${{ env.CACHE_NUMBER }}-${{
-            hashFiles('etc/example-environment.yml') }}
+            hashFiles('conda/environment.yml') }}
       - uses: conda-incubator/setup-miniconda@v2.1.1
         with:
           activate-environment: pointcloudset
