### 2021-09-14 ff257b6
added default value for completly empty datasets
