### 2021-09-23 #3471 @hugoledoux
> @hugoledoux @tgoelles You have my approval :) There are things to improve in the software, but that will always be the case (in any project) but the software is well documented, easy to use and develop for.

@hechth yeah this command is supposed to trigger a bot that takes care of telling people I'm away, but something went wrong... Glad you approved though!
