+++ paper/paper.bib
@@ -142,6 +142,7 @@ url = {https://doi.org/10.5281/zenodo.2556738}
   month     = {May 9-13},
   year      = {2011},
   address   = {Shanghai, China},
+  doi = {10.1109/icra.2011.5980567},
   publisher = {IEEE}
 }
 
