### 2021-09-23 #3471 @hugoledoux
Sorry for the long break, I'm back. I just checked the reviewers' comment and they all accept it so we can proceed with the final steps for approval and publications. Congrats 🎉

I have read the paper and I propose the following small edits:
https://github.com/virtual-vehicle/pointcloudset/pull/31
https://github.com/virtual-vehicle/pointcloudset/pull/32

When this is done, then the following steps are the following:
