+++ CHANGELOG.rst
@@ -6,7 +6,7 @@ The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
 
-Unreleased
+0.3.3 - (2021-09-27)
 -------------
 
 Fixed
