+++ conda/meta.yaml
@@ -1,5 +1,5 @@
 {% set name = "pointcloudset" %}
-{% set version = "0.3.2" %}
+{% set version = "0.3.3" %}
 
 package:
   name: "{{ name|lower }}"
