### 2021-09-27 #3471 @tgoelles
@hugoledoux
The archived version is v0.3.3
The DOI is: 10.5281/zenodo.5529805
https://zenodo.org/record/5529805

I matched the metadata.
So everything should be ready for publication.
