### 2021-09-28 #3471 @hugoledoux
At this point, one editor-in-chief will take it over and might ask you to fix some final things.

But congratulations, your paper is accepted 🎉

Thanks to @RonaldEnsing and @hechth for the reviews
