### 2021-09-28 #3471 @danielskatz
@hugoledoux - for the next time:
1. the paper is not accepted, it's just that acceptance is recommended.  I still need to do some final checks.
2. when the paper is in a branch, you have to give that to the recommend-accept command
