### 2021-09-28 #3471 @danielskatz
@tgoelles - I've made some suggestions for final changes in https://github.com/virtual-vehicle/pointcloudset/pull/33 - please merge this, or let me know what you disagree with, then we can proceed to final acceptance and publishing.
